Interaction Between Creeps & Projectiles

+++ Assets/TowerDefense/Scripts/Turret/ProjectileData.cs
@@ -15,8 +15,8 @@ namespace TowerDefense
     [CreateAssetMenu(fileName = "Base Projectile", menuName = "TowerDefense/Turrets/Projectile")]
     public class ProjectileData : ScriptableObject
     {
-        [Min(1f)]
-        public float damage = 5f;
+        [Min(1)]
+        public int damage = 5;
 
         [Min(0.1f)]
         public float speed = 3f;
